refactor: Image와 ImageDtoAssembler에 흩어진 utils 로직을 ImgaeUtils 파일로 이전
