+++ usecase/src/main/kotlin/me/ogq/ocp/sample/usecase/image/dto/ImageDtoAssembler.kt
@@ -2,9 +2,15 @@ package me.ogq.ocp.sample.usecase.image.dto
 
 import me.ogq.ocp.sample.model.event.ImageEventData
 import me.ogq.ocp.sample.model.image.Image
-import java.io.File
+import me.ogq.ocp.sample.model.image.toUrl
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.stereotype.Component
 
-object ImageDtoAssembler {
+@Component
+class ImageDtoAssembler(
+    @Value("\${const.http-host}")
+    private val httpHost: String
+) {
     fun toDto(image: Image): ImageDto {
         requireNotNull(image.id) { "image.id should not be null" }
 
@@ -16,7 +22,7 @@ object ImageDtoAssembler {
             tags = image.tags,
             imagePath = image.file.path.toString(),
             publicityId = image.publicityRightId.toString(),
-            imageUrl = generateUrl(image.file.path.toString()),
+            imageUrl = image.file.toUrl(httpHost)
         )
     }
 
@@ -31,7 +37,7 @@ object ImageDtoAssembler {
             tags = image.tags,
             imagePath = image.imagePath,
             publicityId = image.publicityRightId.toString(),
-            imageUrl = generateUrl(image.imagePath)
+            imageUrl = image.imagePath.toUrl(httpHost)
         )
     }
 
@@ -48,15 +54,5 @@ object ImageDtoAssembler {
             imagePath = image.file.path.toString(),
             publicityRightId = image.publicityRightId
         )
-    }
-
-    private val httpHost = "http://localhost:8080/"
-    private fun generateUrl(filePath: String): String {
-        fun extractNameFrom(filePath: String): String {
-            return File(filePath).name
-        }
-
-        val fileName = extractNameFrom(filePath)
-        return "$httpHost$fileName"
     }
 }
